new build v.1

+++ src/main/java/com/Group6/WebAppDevGroupProject/Controllers/OrderAdminController.java
@@ -8,48 +8,58 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Optional;
 
 @RestController
 @RequestMapping("orders/api")
 @CrossOrigin
 public class OrderAdminController {
+
     @Autowired
     private OrderService orderService;
 
     @Autowired
     private MenuService menuService;
 
+    // ✅ Get all orders
     @GetMapping("/")
     public List<Order> list() {
         return orderService.getAllOrders();
     }
+
+    // ✅ Get details of a single order
     @GetMapping("/{id}")
-    public Order getOrderDetails(@PathVariable long id) {
+    public Order getOrderDetails(@PathVariable Integer id) {
         return orderService.getOrderById(id);
     }
+
+    // ✅ Get all active orders
     @GetMapping("/active/")
     public List<Order> listActiveOrders() {
         return orderService.getAllActiveOrders();
     }
+
+    // ✅ Get all active orders for a specific user
     @GetMapping("/active/{id}")
-    public List<Order> listActiveUserOrders(@PathVariable long id) {
+    public List<Order> listActiveUserOrders(@PathVariable Integer id) {
         return orderService.getAllActiveOrdersFromUser(id);
     }
+
+    // ✅ Save a new order
     @PostMapping("/")
-    public Order saveOrder(@RequestBody Order ord_)
-    {
+    public Order saveOrder(@RequestBody Order ord_) {
         return orderService.saveOrder(ord_);
     }
+
+    // ✅ Update an existing order
     @PutMapping("/{id_}")
-    public ResponseEntity<Order> updateOrder(@PathVariable long id_, @RequestBody Order ord_)
-    {
+    public ResponseEntity<Order> updateOrder(@PathVariable Integer id_, @RequestBody Order ord_) {
         Order updated_ = orderService.updateOrder(id_, ord_);
         return (updated_ != null) ? ResponseEntity.ok(updated_) : ResponseEntity.notFound().build();
     }
+
+    // ✅ Delete an order
     @DeleteMapping("/{id_}")
-    public ResponseEntity<Void> deleteOrder(@PathVariable long id_)
-    {
+    public ResponseEntity<Void> deleteOrder(@PathVariable Integer id_) {
         orderService.deleteOrder(id_);
         return ResponseEntity.noContent().build();
     }
