+++ src/main/java/com/Group6/WebAppDevGroupProject/Repository/MenuRepository.java
@@ -3,4 +3,4 @@ package com.Group6.WebAppDevGroupProject.Repository;
 import com.Group6.WebAppDevGroupProject.Models.MenuItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<MenuItem, Long> { }
+public interface MenuRepository extends JpaRepository<MenuItem, Integer> { }
